Started the implementation of the lagrangian calculation method, started adding a complementary method that will be used in the calculation of the lagrangian
So far this is only meant to calculate the Lagrangian and not the Euler-Lagrange equations, this calculation is only meant to be implemented once the lagrangian calculation is finished, which is what is in development so far.

the method added only calculates the generalized inertia matrix, which is referred to as D. the main body of the function is implemented but it is still necessary to check if the results are accurate or it needs fixing. During the implementation I had some doubts about how somethings were calculated so the method still needs testing and a theoretical validation