Added comments and simple changes(see body)
Added some comments to the input cell of the notebook, removed some useless and testing cells, so the version kept in the repo is clear.

Somehow I forgot to save before commiting which resulted in not all the changes pointed being commited. I did one more commit to fix this. The other commit also has a reference indicating what happened.